Creates fragments folder, abstracts out of main ViewPageAdapter
