+++ app/src/main/java/com/leyline/namespace/MainActivity.java
@@ -2,19 +2,15 @@ package com.leyline.namespace;
 
 import android.os.Bundle;
 
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
 import androidx.viewpager.widget.ViewPager;
 
 import com.google.android.material.tabs.TabLayout;
-
-import java.text.Normalizer;
-import java.util.ArrayList;
-import java.util.List;
+import com.leyline.namespace.fragments.FormFragment;
+import com.leyline.namespace.fragments.LineFragment;
+import com.leyline.namespace.fragments.PointFragment;
+import com.leyline.namespace.fragments.ShapeFragment;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -62,37 +58,5 @@ public class MainActivity extends AppCompatActivity {
         tabLayout.getTabAt(2).setIcon(R.drawable.shape);
         tabLayout.getTabAt(3).setIcon(R.drawable.form);
 
-    }
-
-    private class ViewPagerAdapter extends FragmentPagerAdapter {
-
-        private List<Fragment> fragments = new ArrayList<>();
-        private List<String> fragmentTitle = new ArrayList<>();
-
-        public ViewPagerAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        public void addFragment(Fragment fragment, String title) {
-            fragments.add(fragment);
-            fragmentTitle.add(title);
-        }
-
-
-        @Override
-        public Fragment getItem(int position) {
-            return fragments.get(position);
-        }
-
-        @Override
-        public int getCount() {
-            return fragments.size();
-        }
-
-        @Nullable
-        @Override
-        public CharSequence getPageTitle(int position) {
-            return fragmentTitle.get(position);
-        }
     }
 }
