+++ app/src/main/java/com/leyline/namespace/ViewPagerAdapter.java
@@ -1,0 +1,41 @@
+package com.leyline.namespace;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class ViewPagerAdapter extends FragmentPagerAdapter {
+
+    private List<Fragment> fragments = new ArrayList<>();
+    private List<String> fragmentTitle = new ArrayList<>();
+
+    public ViewPagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    public void addFragment(Fragment fragment, String title) {
+        fragments.add(fragment);
+        fragmentTitle.add(title);
+    }
+
+
+    @Override
+    public Fragment getItem(int position) {
+        return fragments.get(position);
+    }
+
+    @Override
+    public int getCount() {
+        return fragments.size();
+    }
+
+    @Nullable
+    @Override
+    public CharSequence getPageTitle(int position) {
+        return fragmentTitle.get(position);
+    }
+}
