+++ app/src/main/java/com/leyline/namespace/fragments/FormFragment.java
@@ -1,4 +1,4 @@
-package com.leyline.namespace;
+package com.leyline.namespace.fragments;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.leyline.namespace.R;
 
 /**
  * A simple {@link Fragment} subclass.
