+++ app/src/main/java/com/leyline/namespace/fragments/LineFragment.java
@@ -1,4 +1,4 @@
-package com.leyline.namespace;
+package com.leyline.namespace.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.leyline.namespace.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link PointFragment#newInstance} factory method to
+ * Use the {@link LineFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class PointFragment extends Fragment {
+public class LineFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class PointFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public PointFragment() {
+    public LineFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class PointFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment PointFragment.
+     * @return A new instance of fragment LineFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static PointFragment newInstance(String param1, String param2) {
-        PointFragment fragment = new PointFragment();
+    public static LineFragment newInstance(String param1, String param2) {
+        LineFragment fragment = new LineFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class PointFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_point, container, false);
+        return inflater.inflate(R.layout.fragment_line, container, false);
     }
 }
