+++ app/src/main/java/com/leyline/namespace/fragments/PointFragment.java
@@ -1,4 +1,4 @@
-package com.leyline.namespace;
+package com.leyline.namespace.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.leyline.namespace.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link ShapeFragment#newInstance} factory method to
+ * Use the {@link PointFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class ShapeFragment extends Fragment {
+public class PointFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class ShapeFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public ShapeFragment() {
+    public PointFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class ShapeFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment ShapeFragment.
+     * @return A new instance of fragment PointFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static ShapeFragment newInstance(String param1, String param2) {
-        ShapeFragment fragment = new ShapeFragment();
+    public static PointFragment newInstance(String param1, String param2) {
+        PointFragment fragment = new PointFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class ShapeFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_shape, container, false);
+        return inflater.inflate(R.layout.fragment_point, container, false);
     }
 }
